Extract multihash functions
